+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSets.java
@@ -21,7 +21,7 @@ import net.sourceforge.pmd.lang.ast.RootNode;
 import net.sourceforge.pmd.lang.document.TextFile;
 import net.sourceforge.pmd.lang.rule.internal.RuleApplicator;
 import net.sourceforge.pmd.reporting.FileAnalysisListener;
-import net.sourceforge.pmd.util.log.MessageReporter;
+import net.sourceforge.pmd.util.log.PmdReporter;
 
 /**
  * Grouping of Rules per Language in a RuleSet.
@@ -64,7 +64,7 @@ public class RuleSets {
         this.ruleSets = Collections.singletonList(ruleSet);
     }
 
-    public void initializeRules(LanguageProcessorRegistry lpReg, MessageReporter reporter) {
+    public void initializeRules(LanguageProcessorRegistry lpReg, PmdReporter reporter) {
         // this is abusing the mutability of RuleSet, will go away eventually.
         for (RuleSet rset : ruleSets) {
             for (Iterator<Rule> iterator = rset.getRules().iterator(); iterator.hasNext();) {