+++ pmd-core/src/main/java/net/sourceforge/pmd/internal/util/FileCollectionUtil.java
@@ -21,7 +21,7 @@ import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.util.database.DBMSMetadata;
 import net.sourceforge.pmd.util.database.DBURI;
 import net.sourceforge.pmd.util.database.SourceObject;
-import net.sourceforge.pmd.util.log.MessageReporter;
+import net.sourceforge.pmd.util.log.PmdReporter;
 import net.sourceforge.pmd.util.log.internal.ErrorsAsWarningsReporter;
 
 /**
@@ -57,7 +57,7 @@ public final class FileCollectionUtil {
             // This is to be able to interpret the log (will report 'adding' xxx)
             LOG.debug("Now collecting files to exclude.");
             // errors like "excluded file does not exist" are reported as warnings.
-            MessageReporter mutedLog = new ErrorsAsWarningsReporter(collector.getReporter());
+            PmdReporter mutedLog = new ErrorsAsWarningsReporter(collector.getReporter());
             try (FileCollector excludeCollector = collector.newCollector(mutedLog)) {
 
                 if (configuration.getIgnoreFile() != null) {