+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/document/FileCollector.java
@@ -42,7 +42,7 @@ import net.sourceforge.pmd.lang.Language;
 import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.LanguageVersionDiscoverer;
 import net.sourceforge.pmd.util.AssertionUtil;
-import net.sourceforge.pmd.util.log.MessageReporter;
+import net.sourceforge.pmd.util.log.PmdReporter;
 
 /**
  * Collects files to analyse before a PMD run. This API allows opening
@@ -59,14 +59,14 @@ public final class FileCollector implements AutoCloseable {
     private final List<Closeable> resourcesToClose = new ArrayList<>();
     private Charset charset = StandardCharsets.UTF_8;
     private final LanguageVersionDiscoverer discoverer;
-    private final MessageReporter reporter;
+    private final PmdReporter reporter;
     private final FileId outerFsPath;
     private boolean closed;
     private boolean recursive = true;
 
     // construction
 
-    private FileCollector(LanguageVersionDiscoverer discoverer, MessageReporter reporter, FileId outerFsPath) {
+    private FileCollector(LanguageVersionDiscoverer discoverer, PmdReporter reporter, FileId outerFsPath) {
         this.discoverer = discoverer;
         this.reporter = reporter;
         this.outerFsPath = outerFsPath;
@@ -82,7 +82,7 @@ public final class FileCollector implements AutoCloseable {
      * creating a collector yourself.
      */
     @InternalApi
-    public static FileCollector newCollector(LanguageVersionDiscoverer discoverer, MessageReporter reporter) {
+    public static FileCollector newCollector(LanguageVersionDiscoverer discoverer, PmdReporter reporter) {
         return new FileCollector(discoverer, reporter, null);
     }
 
@@ -90,7 +90,7 @@ public final class FileCollector implements AutoCloseable {
      * Returns a new collector using the configuration except for the logger.
      */
     @InternalApi
-    public FileCollector newCollector(MessageReporter logger) {
+    public FileCollector newCollector(PmdReporter logger) {
         FileCollector fileCollector = new FileCollector(discoverer, logger, null);
         fileCollector.charset = this.charset;
         return fileCollector;
@@ -118,7 +118,7 @@ public final class FileCollector implements AutoCloseable {
      * Returns the reporter for the file collection phase.
      */
     @InternalApi
-    public MessageReporter getReporter() {
+    public PmdReporter getReporter() {
         return reporter;
     }