+++ pmd-core/src/test/java/net/sourceforge/pmd/RulesetFactoryTestBase.java
@@ -30,12 +30,12 @@ import org.slf4j.event.Level;
 import net.sourceforge.pmd.lang.DummyLanguageModule;
 import net.sourceforge.pmd.util.internal.xml.SchemaConstant;
 import net.sourceforge.pmd.util.internal.xml.SchemaConstants;
-import net.sourceforge.pmd.util.log.MessageReporter;
+import net.sourceforge.pmd.util.log.PmdReporter;
 import net.sourceforge.pmd.util.log.internal.SimpleMessageReporter;
 
 public class RulesetFactoryTestBase {
 
-    protected MessageReporter mockReporter;
+    protected PmdReporter mockReporter;
 
     @BeforeEach
     void setup() {